v2.2.0 Added so that schema can be checked if it needs downloading with if-modified header. Better error handling and flow of schema download. Lots of other fixes.

+++ app/src/main/java/com/minder/app/tf2backpack/backend/DownloadSchemaTask.java
@@ -12,18 +12,32 @@ import com.minder.app.tf2backpack.Util;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.MalformedURLException;
-import java.util.Date;
 
 /**
  * Created by Patrik on 2018-09-15.
  */
 
-public class DownloadSchemaOverviewTask extends AsyncTask<Void, Float, Void> {
+public class DownloadSchemaTask extends AsyncTask<Void, Float, DownloadSchemaTask.Result> {
 
     public interface ProgressListener {
         void onProgress(float t);
-        void onComplete(Date dataLastModified);
+        void onComplete(long dataLastModified);
         void onError(Exception error);
+    }
+
+    public static class Result {
+        final long lastModifiedTimestamp;
+        final Exception error;
+
+        Result(long lastModifiedTimestamp) {
+            this.lastModifiedTimestamp = lastModifiedTimestamp;
+            this.error = null;
+        }
+
+        Result(Exception error) {
+            this.lastModifiedTimestamp = -1;
+            this.error = error;
+        }
     }
 
     private static final float OVERVIEW_PERCENT = 0.3f;
@@ -31,17 +45,23 @@ public class DownloadSchemaOverviewTask extends AsyncTask<Void, Float, Void> {
 
     private Context context;
     private ProgressListener listener;
+    private long ifModifiedSinceTimestamp;
 
-    public DownloadSchemaOverviewTask(Context context, ProgressListener listener) {
+    public DownloadSchemaTask(Context context, long ifModifiedSinceTimestamp, ProgressListener listener) {
         this.context = context;
         this.listener = listener;
+        this.ifModifiedSinceTimestamp = ifModifiedSinceTimestamp;
     }
 
-    public void downloadParseAndSaveSchemaFilesToDb(SQLiteDatabase sqlDb, int start, final int parsedItems) throws Exception {
+    public boolean downloadParseAndSaveSchemaFilesToDb(SQLiteDatabase sqlDb, long ifModifiedSinceTimestamp, int start, final int parsedItems) throws Exception {
 		HttpConnection connection =
 				HttpConnection.string("http://api.steampowered.com/IEconItems_440/GetSchemaItems/v1/?key=" +
                         ApiKey.get() + "&format=json&language=en" + (start >= 0 ? "&start=" + start : ""));
-        
+
+        if (ifModifiedSinceTimestamp > 0) {
+            connection.setIfModifiedSince(ifModifiedSinceTimestamp);
+        }
+
         // Function with limit at 1 and will be about 90% on the way when we have parsed 5000 items
         // which is the close to the total amount of items in the game at the time of writing.
         // This ensures we do not surpass 1 if more items are added and the app is not changed.
@@ -76,18 +96,35 @@ public class DownloadSchemaOverviewTask extends AsyncTask<Void, Float, Void> {
         });
 
 		if (inputStream != null) {
-			GameSchemeParser gs;
-			try {
-				gs = new GameSchemeParser(inputStream, sqlDb);
-				publishProgress(totalProgressSoFar + ourSegmentOfTotal);
-			} finally {
-				try {
-					inputStream.close();
-				} catch (IOException e) {}
-			}
-			if (gs.getNextStart() >= 0) {
-				downloadParseAndSaveSchemaFilesToDb(sqlDb, gs.getNextStart(), parsedItems + gs.getParsedItems());
-			}
+            boolean itemsModified = true;
+            try {
+                itemsModified = connection.getConnection().getResponseCode() != 304;
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
+            if (itemsModified) {
+                GameSchemeParser gs;
+                try {
+                    gs = new GameSchemeParser(inputStream, sqlDb);
+                    publishProgress(totalProgressSoFar + ourSegmentOfTotal);
+                } finally {
+                    try {
+                        inputStream.close();
+                    } catch (IOException e) {
+                    }
+                }
+                if (gs.getNextStart() >= 0) {
+                    return downloadParseAndSaveSchemaFilesToDb(sqlDb, -1, gs.getNextStart(), parsedItems + gs.getParsedItems());
+                }
+                return true;
+            }
+            else {
+                try {
+                    inputStream.close();
+                } catch (IOException e) {}
+                return false;
+            }
 		}
 		else {
 			throw connection.getException();
@@ -95,7 +132,7 @@ public class DownloadSchemaOverviewTask extends AsyncTask<Void, Float, Void> {
 	}
 
     @Override
-    protected Void doInBackground(Void... voids) {
+    protected DownloadSchemaTask.Result doInBackground(Void... voids) {
         HttpConnection connection;
         try {
             connection = HttpConnection.string("http://api.steampowered.com/IEconItems_440/GetSchemaOverview/v1/?key=" +
@@ -104,6 +141,10 @@ public class DownloadSchemaOverviewTask extends AsyncTask<Void, Float, Void> {
         catch (MalformedURLException e) {
             e.printStackTrace();
             return null;
+        }
+
+        if (ifModifiedSinceTimestamp > 0) {
+            connection.setIfModifiedSince(ifModifiedSinceTimestamp);
         }
 
         final InputStream inputStream = connection.executeStream(new HttpConnection.DownloadProgressListener() {
@@ -128,11 +169,19 @@ public class DownloadSchemaOverviewTask extends AsyncTask<Void, Float, Void> {
         });
 
         if (inputStream != null) {
+            boolean schemaOverviewModified = true;
+            try {
+                schemaOverviewModified = connection.getConnection().getResponseCode() != 304;
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            final boolean schemaOverviewModifiedFinal = schemaOverviewModified;
+            long lastModified = connection.getConnection().getLastModified();
             Log.i(Util.GetTag(), "Saving to database...");
             long start = System.currentTimeMillis();
 
-            DataBaseHelper.runWithWritableDb(context, new DataBaseHelper.RunWithWritableDb() {
-                public void run(SQLiteDatabase sqlDb) {
+            Exception error = DataBaseHelper.runWithWritableDb(context, new DataBaseHelper.RunWithWritableDb<Exception>() {
+                public Exception run(SQLiteDatabase sqlDb) {
                     sqlDb.beginTransaction();
                     try {
                         // Cleanup old data from db
@@ -140,51 +189,68 @@ public class DownloadSchemaOverviewTask extends AsyncTask<Void, Float, Void> {
                             // removes everything from database, else we would get duplicates
                             sqlDb.delete("items", null, null);
                             sqlDb.delete("item_attributes", null, null);
-                            sqlDb.delete("attributes", null, null);
-                            sqlDb.delete("item_attributes", null, null);
-                            sqlDb.delete("particles", null, null);
-                            sqlDb.delete("strange_score_types", null, null);
-                            sqlDb.delete("item_qualities", null, null);
 
-                            // we need to fetch table names
-                            final Cursor c = sqlDb.rawQuery("SELECT type_name FROM strange_item_levels", null);
-                            while (c.moveToNext()) {
-                                // delete each table for item levels
-                                sqlDb.execSQL("DROP TABLE " + c.getString(0));
+                            if (schemaOverviewModifiedFinal) {
+                                sqlDb.delete("attributes", null, null);
+                                sqlDb.delete("particles", null, null);
+                                sqlDb.delete("strange_score_types", null, null);
+                                sqlDb.delete("item_qualities", null, null);
+
+                                // we need to fetch table names
+                                final Cursor c = sqlDb.rawQuery("SELECT type_name FROM strange_item_levels", null);
+                                while (c.moveToNext()) {
+                                    // delete each table for item levels
+                                    sqlDb.execSQL("DROP TABLE " + c.getString(0));
+                                }
+                                c.close();
+
+
+                                sqlDb.delete("strange_item_levels", null, null);
                             }
-                            c.close();
-
-                            sqlDb.delete("strange_item_levels", null, null);
                         }
 
-                        new GameSchemaOverviewParser(inputStream, sqlDb);
+                        if (schemaOverviewModifiedFinal) {
+                            new GameSchemaOverviewParser(inputStream, sqlDb);
+                        }
                         try {
                             inputStream.close();
                         } catch (IOException e) { }
                         
                         publishProgress(OVERVIEW_PERCENT);
-                        downloadParseAndSaveSchemaFilesToDb(sqlDb, -1, 0);
-                        sqlDb.setTransactionSuccessful();
-                    } catch (Exception e1) {
-                        e1.printStackTrace();
+                        // If the overView data has changed, then we force update the rest to be safe.
+                        boolean hasChanges = downloadParseAndSaveSchemaFilesToDb(
+                                sqlDb,
+                                schemaOverviewModifiedFinal ? -1 : ifModifiedSinceTimestamp,
+                                -1,
+                                0);
+                        if (hasChanges) {
+                            sqlDb.setTransactionSuccessful();
+                        }
+                    } catch (Exception error) {
+                        return error;
                     } finally {
                         try {
                             inputStream.close();
                         } catch (IOException e) { }
                         sqlDb.endTransaction();
                     }
-                    
+
+                    return null;
                 }
             });
 
-            publishProgress(1.0f);
-            Log.i(Util.GetTag(), "Save to database finished - Time: " + (System.currentTimeMillis() - start) + " ms");
+            if (error == null) {
+                publishProgress(1.0f);
+                Log.i(Util.GetTag(), "Save to database finished - Time: " + (System.currentTimeMillis() - start) + " ms");
+                return new Result(lastModified);
+            }
+            else {
+                return new Result(error);
+            }
         }
         else {
-            connection.getException().printStackTrace();
+            return new Result(connection.getException());
         }
-
-        return null;
     }
 
     @Override
@@ -194,7 +260,11 @@ public class DownloadSchemaOverviewTask extends AsyncTask<Void, Float, Void> {
 	}
 
     @Override
-    protected void onPostExecute(Void aVoid) {
-        listener.onComplete(new Date());
+    protected void onPostExecute(DownloadSchemaTask.Result result) {
+        if (result.error == null) {
+            listener.onComplete(result.lastModifiedTimestamp);
+        } else {
+            listener.onError(result.error);
+        }
     }
 }
